Todo actions converted to typescript
